disable set-user.sh which causes permission denied
I think the underlying cause is that the Dockerfile sets the user to 1001, but the
startup command for the container includes set-user.sh, which copies the /etc/passwd
file into place from a filtered template.

I don't see how this could be Openshift-compatible, since the image would have to start
as the root user in order to have these kinds of permissions.

See also: MMENG-431